Consolidating disparate namespaces
With this commit, I'm replacing the WorkConverter class with a
WorkConverter module. This also means I'm pushing the metadata
definitions down another level.
