+++ app/conversions/sipity/conversions/to_rof_hash/work_converters/etd_converter.rb
@@ -1,8 +1,8 @@
-require 'sipity/conversions/to_rof_hash/specific_work_converters/abstract_converter'
+require 'sipity/conversions/to_rof_hash/work_converters/abstract_converter'
 module Sipity
   module Conversions
     module ToRofHash
-      module SpecificWorkConverters
+      module WorkConverters
         # Responsible for exposing the custom metadata mapping that occurs for an ETD
         class EtdConverter < AbstractConverter
           def af_model
