+++ spec/conversions/sipity/conversions/to_rof_hash/work_converter_spec.rb
@@ -2,27 +2,65 @@ require 'spec_helper'
 require 'sipity/conversions/to_rof_hash/work_converter'
 
 module Sipity
-  RSpec.describe Conversions::ToRofHash::WorkConverter do
-    let(:work) do
-      Sipity::Models::Work.new(
-        id: 'abcd-ef',
-        work_type: 'doctoral_dissertation',
-        collaborators: [collaborator],
-        access_right: access_right
-      )
+  module Conversions
+    module ToRofHash
+      RSpec.describe WorkConverter do
+        let(:work) { double('Work') }
+        let(:converting_instance) { double(to_hash: true) }
+        subject { described_class }
+
+        context '.call' do
+          it 'should convert the work into a hash' do
+            allow(described_class).to receive(:find_and_initialize).and_return(converting_instance)
+            subject.call(work: work)
+            expect(converting_instance).to have_received(:to_hash)
+          end
+        end
+
+        its(:default_repository) { is_expected.to respond_to(:scope_users_for_entity_and_roles) }
+
+        context '.find_and_initialize' do
+          let(:repository) { QueryRepositoryInterface.new }
+          subject { described_class.find_and_initialize(work: work, repository: repository) }
+          context 'for a Models::WorkType::DOCTORAL_DISSERTATION' do
+            let(:work) { Sipity::Models::Work.new(work_type: Models::WorkType::DOCTORAL_DISSERTATION) }
+            it { is_expected.to be_a(WorkConverters::EtdConverter) }
+          end
+          context 'for a Models::WorkType::MASTER_THESIS' do
+            let(:work) { Sipity::Models::Work.new(work_type: Models::WorkType::MASTER_THESIS) }
+            it { is_expected.to be_a(WorkConverters::EtdConverter) }
+          end
+
+          context 'for a Models::WorkType::ULRA_SUBMISSION' do
+            let(:work) { Sipity::Models::Work.new(work_type: Models::WorkType::ULRA_SUBMISSION) }
+            {
+              '10000 Level' => WorkConverters::UlraDocumentConverter,
+              "20000–40000 Level" => WorkConverters::UlraDocumentConverter,
+              "Honors Thesis" => WorkConverters::UlraDocumentConverter,
+              "Capstone Project" => WorkConverters::UlraDocumentConverter,
+              'Senior Thesis' => WorkConverters::UlraSeniorThesisConverter
+            }.each_pair do |award_category, converter_class|
+              context "with award category #{award_category.inspect}" do
+                before { allow(repository).to receive(:work_attribute_values_for).and_return(award_category) }
+                it { is_expected.to be_a(converter_class) }
+              end
+            end
+            context "with an unknown award category" do
+              before { allow(repository).to receive(:work_attribute_values_for).and_return('Best Haircut') }
+              it 'should raise an Exceptions::FailedToInitializeWorkConverterError' do
+                expect { subject }.to raise_error(Exceptions::FailedToInitializeWorkConverterError)
+              end
+            end
+          end
+          context 'for an unhandled work type' do
+            let(:work) { double(work_type: 'coalmine') }
+            it 'should raise an Exceptions::FailedToInitializeWorkConverterError' do
+              expect { subject }.to raise_error(Exceptions::FailedToInitializeWorkConverterError)
+            end
+          end
+        end
+
+      end
     end
-    let(:access_right) { Sipity::Models::AccessRight.new(access_right_code: Sipity::Models::AccessRight::OPEN_ACCESS) }
-    let(:collaborator) { Sipity::Models::Collaborator.new(role: 'Research Director', name: 'Alexander Hamilton') }
-    let(:repository) { Sipity::QueryRepositoryInterface.new }
-    subject { described_class.new(work: work, repository: repository) }
-
-    it 'exposes .call as a convenience method' do
-      expect_any_instance_of(described_class).to receive(:call)
-      described_class.call(work: work)
-    end
-
-    its(:default_repository) { is_expected.to respond_to(:scope_users_for_entity_and_roles) }
-
-    its(:call) { is_expected.to be_a(Hash) }
   end
 end
